requirements.txt
selenium
pillow